% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined}
\alias{combined}
\title{Combined data set of geochemicals and OTU abundance}
\format{A data frame with 7 rows and 10 variables:
\describe{
  \item{Cruise}{Numerical identifier of individual cruises. Units: Numeric
  string}
  \item{Depth}{Sampling depth. Units: Meters}
  \item{OTU0001}{Operational taxonomic unit 1}
  \item{OTU0002}{Operational taxonomic unit 2}
  \item{OTU0003}{Operational taxonomic unit 3}
  \item{OTU0004}{Operational taxonomic unit 4}
  \item{NO3}{Nitrate (Bran Luebbe AutoAnalyser - colorimetric). Units:
  Micromolar}
  \item{Mean_NO2}{Average concentration of Nitrite for each depth
  (colorimetric assay). Units: Micromolar}
  \item{Mean_N2O}{Average concentration of Nitrous oxide (headspace or
  automated purge-and-trap). Units: Nanomolar}
  \item{Mean_NH4}{Average concentration of Ammonium for each depth
  (fluorometric assay). Units: Micromolar}
}}
\source{
For a brief introduction to the data, see Hallam SJ \emph{et al.}
  2017. Monitoring microbial responses to ocean deoxygenation in a model
  oxygen minimum zone. Sci Data 4: 170158. doi:
  \href{https://www.nature.com/articles/sdata2017158}{10.1038/sdata.2017.158}.
   More detailed information on the geochemical data can be found in
  Torres-Beltrán M \emph{et al.} 2017. A compendium of geochemical
  information from the Saanich Inlet water column. Sci Data 4: 170159. doi:
  \href{https://www.nature.com/articles/sdata2017159}{10.1038/sdata.2017.159}
  (\href{https://datadryad.org/stash/dataset/doi:10.5061/dryad.nh035}{dataset}).
  More detailed information on the mutli-omic microbial data can be found in
  Hawley AK *et al*. 2017. A compendium of multi-omic sequence information
  from the Saanich Inlet water column. Sci Data 4: 170160. doi:
  \href{https://www.nature.com/articles/sdata2017160}{10.1038/sdata.2017.160}.
}
\usage{
combined
}
\description{
A subset of the \link[=geochemicals]{geochemical} and
\link[=normalized_OTUs]{normalized OTU abundance}) data sets at 7 depths for
Saanich Inlet Cruise 72.
}
\keyword{datasets}
