% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educer.R
\docType{package}
\name{educer}
\alias{educer}
\title{educer: A collection of tutorials on using R in Data Science.}
\description{
The educer package was created by the Experiential Data Science for
Undergraduate Cross-disciplinary Education (EDUCE) initiative at the
University of British Columbia. It contains learnr tutorials introducing
basic data processing and wrangling operation as a pre-class activity.
}
\section{Tutorials}{

Tutorials are started using
\code{learnr::run_tutorial(name, package = "educer")}, where the following
\code{name}s start the corresponding tutorial:

\code{r_and_rstudio_basic}: Introduction to R and RStudio

\code{tidyverse_intermediate}: Working with data

\code{statistics_intermediate}: Introduction to *t*-test, ANOVA and linear
regression

\code{ggplot_basic}: Basic plotting

\code{ggplot_intermediate}: Advanced plotting
}

