% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tutorials}
\alias{list_tutorials}
\title{List tutorials in the educer package}
\usage{
list_tutorials()
}
\description{
This wrapper lists learnr tutorials specifically in the educer package.
}
\seealso{
\code{\link[learnr]{available_tutorials}()}
}
