% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Measurements of geochemicals and OTU abundance at 7 depths}
\format{A data frame with 7 rows and 10 variables:
\describe{
  \item{sample}{Samples collected at different depths, designated by _#}
  \item{Otu0001}{Operational taxonomic units 1 to 4}
  \item{Otu0002}{Operational taxonomic units 1 to 4}
  \item{Otu0003}{Operational taxonomic units 1 to 4}
  \item{Otu0004}{Operational taxonomic units 1 to 4}
  \item{Depth_m}{Depth in meters}
  \item{NO3_uM}{Nitrate concentration in micromolar}
  \item{NO2_uM}{Nitrite concentration in micromolar}
  \item{N2O_uM}{Nitrous oxide concentration in micromolar}
  \item{NH4_uM}{Ammonium concentration in micromolar}
}}
\source{
Geochemical \url{https://www.nature.com/articles/sdata2017159} and
Multi-omics \url{https://www.nature.com/articles/sdata2017160}
}
\usage{
meta
}
\description{
A subset of the Saanich geochemical and OTU data sets. Needs to be replaced
}
\keyword{datasets}
