% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OTUs}
\alias{OTUs}
\title{Abundance of OTUs for Saanich Inlet Cruise 72}
\format{A data frame with 30,576 rows and 4 variables:
\describe{
  \item{Cruise}{Numerical identifier of individual cruises. Units: Numeric
  string}
  \item{Depth}{Sampling depth. Units: Meters}
  \item{OTU}{Numerical identifier of OTU. Units: Numeric string}
  \item{Count}{Count of OTU sequences. Units: Numeric string}
}}
\source{
For a brief introduction to the data, see Hallam SJ \emph{et al.}
  2017. Monitoring microbial responses to ocean deoxygenation in a model
  oxygen minimum zone. Sci Data 4: 170158
  \href{https://www.nature.com/articles/sdata2017158}{doi:10.1038/sdata.2017.158}.
   More detailed information on the mutli-omic microbial data can be found in
  Hawley AK *et al*. 2017. A compendium of multi-omic sequence information
  from the Saanich Inlet water column. Sci Data 4: 170160.
  \href{https://www.nature.com/articles/sdata2017160}{doi:10.1038/sdata.2017.160}.
}
\usage{
OTUs
}
\description{
Saanich Inlet, a seasonally anoxic fjord on the coast of Vancouver Island
British Columbia, is a model ecosystem for monitoring microbial community
structure, function and dynamics in relation to changing levels of water
column oxygen-deficiency extensible to coastal and open ocean OMZs. This data
set contains cruise 72 of monthly time series observations in Saanich Inlet
encompassing environmental sequence information. Sequencing data was
processed using the mothur_pipeline R Markdown file to bin the data into
operational taxonomic units and determine their abundance (alternatively, use
the \link[=normalized_OTUs]{normalized data set}).
% Find way to link Rmd file
}
\keyword{datasets}
