% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_up_resources}
\alias{set_up_resources}
\title{Set up paths to resources used in learnr tutorials}
\usage{
set_up_resources()
}
\description{
This helper function provides the paths for any resources used in learnr
tutorials. During package development, each <tutorial> is located in
\code{inst/tutorials/<tutorial>} and resources in \code{inst/resources}
(relative to the root of the package). After installation, <tutorial> is in
\code{tutorials/<tutorial>} and resources in \code{resources} (relative to
the root of the installed package). \code{shiny::addResourcePath} creates a
prefix consisting of the name of each directory in \code{resources} and the
file path of that directory. For example, the prefix to \code{/images}) is
connected to \code{resources/images}, etc.
}
\seealso{
\code{\link[shiny]{addResourcePath}()}
}
\keyword{internal}
