% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start_tutorial}
\alias{start_tutorial}
\title{Start tutorial in the educer package}
\usage{
start_tutorial(name)
}
\arguments{
\item{name}{A string giving the name of the educer tutorial}
}
\description{
This wrapper runs learnr tutorials specifically in the educer package.
}
\seealso{
\code{\link[learnr]{run_tutorial}()}
}
